==> governance/03-governance-making-index.Rmd <==
---
title: "Governance - Making the Index"
author: "Olivia Beck"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
options(width = 100)
```

```{r, message= FALSE}
load("governance/governance-design-matrix.Rdata")
library(dplyr)
library(purrr)
library(stringr)
library(tidyr)
library(kableExtra)
library(data.table)
library(polycor)
library(ggcorrplot)
library(psych)
library(maps)

#devtools::install_github("UrbanInstitute/nccsdata")
library(nccsdata)

#devtools::install_github("UrbanInstitute/urbnmapr")
library(urbnmapr)

```

## Summary of Variables

### Variables Used in Index

All variables are coded as 1 (good governance practice) or 0 (bad governance practice)

+------------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
| Variable Name    | Variable Meaning/Coding                                                                                                                                                                                                           |
+==================+===================================================================================================================================================================================================================================+
| P4_LINE_12A      | Did the org obtain separate, independent audited financial statements? 1 - yes, 0 - no                                                                                                                                            |
+------------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
| P4_LINE_28       | 1- Answered "no" to questions P4_LINE_28A, P4_LINE_28B, **AND** P4_LINE_28                                                                                                                                                        |
|                  |                                                                                                                                                                                                                                   |
|                  | 0 - Answered yes to any of these questions                                                                                                                                                                                        |
|                  |                                                                                                                                                                                                                                   |
|                  | Essentially 1 if no business was done with a family member and 0 otherwise.                                                                                                                                                       |
+------------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
| P4_LINE_29_30    | 1- (Answered no to P4_LINE_29 ANDP4_Line_29) or (Answered yes to (P4_LINE_29 ORP4_Line_29) AND answered yes to Schedule M question 31 )                                                                                           |
|                  |                                                                                                                                                                                                                                   |
|                  | 0 - Answered yes to (P4_LINE_29 ORP4_Line_29) AND answered no to Schedule M question 31                                                                                                                                           |
|                  |                                                                                                                                                                                                                                   |
|                  | i.e. 1 if either (the org did not have more than \$25K in cash contribution AND did not have receive any art/historical treasures/etc.) OR (they do have those things AND they do have a written policy about how to handle them) |
+------------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
| P6_LINE_1        | 1 - The proportion of independent members on the board is at least 50%                                                                                                                                                            |
|                  |                                                                                                                                                                                                                                   |
|                  | 0 - The proportion of independent members on the board is less than 50%                                                                                                                                                           |
+------------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
| P6_LINE_2        | 1- **No** officer, director, trustee, or key employee had a family relationship or a business relationship with any other officer, director, trustee, or key employee                                                             |
|                  |                                                                                                                                                                                                                                   |
|                  | 0- **An** officer, director, trustee, or key employee had a family relationship or a business relationship with any other officer, director, trustee, or key employee                                                             |
+------------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
| P6_LINE_3        | 1- The org **did not** use a management company or other person to perform any management duties                                                                                                                                  |
|                  |                                                                                                                                                                                                                                   |
|                  | 0 - The org **did** use a management company or other person to perform any management duties                                                                                                                                     |
+------------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
| P6_LINE_8A       | 1 - The meetings by the governing body **are** documented                                                                                                                                                                         |
|                  |                                                                                                                                                                                                                                   |
|                  | 0 - The meetings by the governing body **are** **not** documented                                                                                                                                                                 |
+------------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
| P6_LINE_12_13_14 | 1- The answers to P6_Line_12, 13, and 14 are all yes                                                                                                                                                                              |
|                  |                                                                                                                                                                                                                                   |
|                  | 0- otherwise                                                                                                                                                                                                                      |
|                  |                                                                                                                                                                                                                                   |
|                  | i.e. 1 if the organization has a written conflict of interest policy, **AND** a written whistleblower policy, **AND** a written retention and destruction policy.                                                                 |
|                  |                                                                                                                                                                                                                                   |
|                  | These 3 are combined because seperated they have \>0.90 correlation. And all of these questions get at the same type of governance - do you have written policies for when things go bad?                                         |
+------------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
| P6_LINE_11A      | 1- A copy of Form 990 is provided to all members before filing the form                                                                                                                                                           |
|                  |                                                                                                                                                                                                                                   |
|                  | 0 - otherwise                                                                                                                                                                                                                     |
+------------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
| P6_LINE_15A      | 1- The org has a process for determining compensation of the CEO, executive director, or top management officials including a review and approval by independent persons, comparability data, and deliberation decision           |
|                  |                                                                                                                                                                                                                                   |
|                  | 0 - otherwise                                                                                                                                                                                                                     |
+------------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
| P6_LINE_18       | 1 - Forms 1023 (or 1024), 990, 990-T available for public inspection on own website                                                                                                                                               |
|                  |                                                                                                                                                                                                                                   |
|                  | 0 - These forms were listed on another's website, available upon request, or "other"                                                                                                                                              |
+------------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
| P12_LINE_1       | Did the org use accrual method for accouting? 1 - yes, 0- no                                                                                                                                                                      |
+------------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+

### Variables Not Used in Index

+--------------------------+----------------------------------------------------+
| Variable Name            | Reason why it is not used                          |
+==========================+====================================================+
| P4_LINE_12B              | Perfectly correlated with P4_Line_12A              |
+--------------------------+----------------------------------------------------+
| P6_LINE_4                | Removed, neither good nor bad                      |
+--------------------------+----------------------------------------------------+
| P6_Line_5                | Less than 3% in 0 category                         |
+--------------------------+----------------------------------------------------+
| P6_Line_6                | Removed, neither good nor bad                      |
+--------------------------+----------------------------------------------------+
| P6_LINE_7A, P6_LINE_7B   | Causing singularity issues                         |
+--------------------------+----------------------------------------------------+
| P6_Line_8B               | Perfectly correlated with P6_Line_8A               |
+--------------------------+----------------------------------------------------+
| P6_Line_12B, P6_LINE_12C | Folded into question 12A, see make-govern-data.Rmd |
+--------------------------+----------------------------------------------------+
| P6_LINE_15A              | Perfectly correlated with P6_Line_15A              |
+--------------------------+----------------------------------------------------+
| P6_LINE_16A, P6_LINE_16B | Less than 3% in 0 category                         |
+--------------------------+----------------------------------------------------+
| P12_LINE_2A, P12_LINE_2C | Less than 3% in 1 category                         |
+--------------------------+----------------------------------------------------+
| P12_LINE_2B              | Perfectly correlated with P12_Line_2A              |
+--------------------------+----------------------------------------------------+

## Doing the Factor Analysis

### Importing data and wrangling

```{r}
#Load in design matrix called "features2" 
load("governance/governance-design-matrix.Rdata")

#Select correct variabels 
keep_vars <- c("P12_LINE_1", 
               "P4_LINE_12", "P4_LINE_29_30", 
               "P6_LINE_1", "P6_LINE_2", "P6_LINE_3", "P6_LINE_8A",
               "P6_LINE_11A", "P6_LINE_12_13_14", "P6_LINE_15A", "P6_LINE_18" )

#update design matrix 
design_mat2 <- as.data.frame(features2)
design_mat2 <- sapply(design_mat2[, keep_vars], as.factor)

#corelations matrix - pearson corelations
corr_mat2 <- hetcor(design_mat2, ML = TRUE)

#plot corrleations
ggcorrplot(corr_mat2$correlations,
           hc.order = TRUE ) +
  ggtitle("Correlation Matrix") +
  labs(subtitle = "Variables used in factor analysis")


# corelations matrix - polychoric 
poly_cor2 = polychoric(features2)
rho2 = poly_cor2$rho

#scree plot 
fa.parallel(rho2, 
            fm="pa", 
            fa="fa", 
            main = "Scree Plot",
            cor = "poly",
            nfactors = 1,
            n.obs = nrow(design_mat2))

# EFA with 6 dimensions 
model2.6 = fa(features2, nfactor=6, cor="poly", fm="mle", rotate = "equamax", 
              scores = "Thurstone")


```

### Let's look at the model and diagnostics

```{r}
summary(model2.6)

#loadings and variance 
loadings <- model2.6$loadings
loadings

#loadings diagram
fa.diagram(model2.6)
scores2.6 <- model2.6$scores

#How well does the factor model reproduce the correlation matrix?
model2.6$fit

#how well are the off diagonal elements reproduced?
model2.6$fit.off

#pvalue
model2.6$PVAL

```

## Adding Factor Scores to Design Matrix

```{r}

scores <- factor.scores(features2, 
                        model2.6, 
                        rho = polychoric(features2)$rho,
                        method = "Thurstone")


all.dat <- cbind(features2, scores$scores)
all.dat <- as.data.frame(all.dat)

all.dat$total.score <- rowSums(scores$scores)


#correlation matrix of weights 
#highly uncorrelated so this is fine
# would be perfectly not corrleated if we used pearson correlations
ggcorrplot(scores$r.scores,
           hc.order = TRUE ) +
  ggtitle("Correlation Matrix of Factors")

```

### Exploring Results

```{r}
all.dat %>% 
  select(ML1, ML2, ML3, ML4, ML5, ML6) %>% 
  pivot_longer(everything(), names_to = "factors", values_to = "vals") %>%
  ggplot() +
  geom_histogram(aes(x = vals)) +
  facet_wrap(~ factors) +
  ggtitle("Scores by Factor")

## Factor 1 (ML2) - p4_line1, p12_line1, p6_line12
all.dat %>% 
  select(P4_LINE_12, P12_LINE_1, P6_LINE_12_13_14, ML2) %>% 
  pivot_longer(!ML2, names_to = "vars", values_to = "values") %>% 
  mutate(values = as.factor(values),
         vars = as.factor(vars)) %>%
  ggplot(aes(x = ML2)) +
  geom_histogram() +
  facet_grid(vars ~ values) +
  ggtitle("Scores within Factor ML2",
          subtitle = "By Originial Variable and Value")

## Factor 2 (ML1) - P6_LINE_8A
all.dat %>% 
  select(P6_LINE_8A, ML1) %>% 
  pivot_longer(!ML1, names_to = "vars", values_to = "values") %>% 
  mutate(values = as.factor(values),
         vars = as.factor(vars)) %>%
  ggplot(aes(x = ML1)) +
  geom_histogram() +
  facet_grid(vars ~ values) +
  ggtitle("Scores within Factor ML1",
          subtitle = "By Originial Variable and Value")

## Factor 3 (ML4) - P6_LINE_15A, P6_LINE_1
all.dat %>% 
  select(P6_LINE_15A, P6_LINE_1, ML4) %>% 
  pivot_longer(!ML4, names_to = "vars", values_to = "values") %>% 
  mutate(values = as.factor(values),
         vars = as.factor(vars)) %>%
  ggplot(aes(x = ML4)) +
  geom_histogram() +
  facet_grid(vars ~ values) +
  ggtitle("Scores within Factor ML4",
          subtitle = "By Originial Variable and Value")

## Factor 4 (ML3) - P6_line_3
all.dat %>% 
  select(P6_LINE_3, ML3) %>% 
  pivot_longer(!ML3, names_to = "vars", values_to = "values") %>% 
  mutate(values = as.factor(values),
         vars = as.factor(vars)) %>%
  ggplot(aes(x = ML3)) +
  geom_histogram() +
  facet_grid(vars ~ values) +
  ggtitle("Scores within Factor ML3",
          subtitle = "By Originial Variable and Value")

## Factor 5 (ML6) - P6_line_11A, P6_LINE_18
all.dat %>% 
  select(P6_LINE_11A, P6_LINE_18, ML6) %>% 
  pivot_longer(!ML6, names_to = "vars", values_to = "values") %>% 
  mutate(values = as.factor(values),
         vars = as.factor(vars)) %>%
  ggplot(aes(x = ML6)) +
  geom_histogram() +
  facet_grid(vars ~ values) +
  ggtitle("Scores within Factor ML6",
          subtitle = "By Originial Variable and Value")

## Factor 6 (ML5) - P6_line_2, P4_line_28, P4_Line_29_30
all.dat %>% 
  select(P6_LINE_2, P4_LINE_28, P4_LINE_29_30, ML5) %>% 
  pivot_longer(!ML5, names_to = "vars", values_to = "values") %>% 
  mutate(values = as.factor(values),
         vars = as.factor(vars)) %>%
  ggplot(aes(x = ML5)) +
  geom_histogram() +
  facet_grid(vars ~ values) +
  ggtitle("Scores within Factor ML5",
          subtitle = "By Originial Variable and Value")


```

## Showing How Scores are Calculated

There are 12 variables, 6 factors, and n$\approx$ 64k observations. Let,

-   $D$ = the Design matrix $\in \mathbb{R}^{n x 12}$

-   $D_s$ = Design matrix with columns centered at 0 and scaled to have 1 sd $\in \mathbb{R}^{n x 12}$

-   $C$ = observed correlation matrix $\in \mathbb{R}^{12 x 12}$. In our example, this is the polychoric correlation.

-   $L$ = factor loading matrix $\in \mathbb{R}^{12 x 6}$

-   $W = C^{-1}L$ = matrix of weights ($\beta$'s in the regression model) $\in \mathbb{R}^{12 x 6}$

-   $S = D_s W$ = matrix of scores $\in \mathbb{R}^{n x 6}$

Thus if $D_s^*$ is a new (scaled) design matrix using the same scaling mean and sd as $D_S$, then $S^* = D_s^* W$ is the matrix of new scores.

```{r}
data.scaled <- scale(features2)
D <- data.scaled
C <- model2.6$r  # =  polychoric(features2)$rho
W <- scores$weights
L <- model2.6$loadings #same as model2.6$Structure
S <- scores$scores
n <- nrow(features2)

#S = DW so lets check DW - S
sum(D %*% W - S)

#W = C^(-1) L so lets check  C^(-1) L - W
sum(solve(C) %*% L - W)
```

### If we were to get new data, this is how we would get the new scores

```{r}
scaled.mean <- attr(data.scaled, "scaled:center")
scaled.sd <- attr(data.scaled, "scaled:scale")

N <- 100
new.data <- matrix(NA, ncol = ncol(features2), nrow = N)
colnames(new.data) <- colnames(features2)
for(i in 1:ncol(features2)){
  new.data[, i] <- rbinom(N, 1, mean(features2[, i]))
}

new.scores<- factor.scores(new.data, 
                        model2.6, 
                        rho = polychoric(features2)$rho,
                        method = "Thurstone")

features2$new.scores <- new.scores$scores

### plot of distributions of scores 
hist(rowSums(new.scores$scores))

```

## Exploring Distributions of Scores

```{r}
#grabbing other "demographic" information about the orgs 
dat.2018 <- get_data(dsname = "core",
                 time = "2018")

#bind ID's and values
all.dat <- cbind(features2_ID, all.dat) 

remove.ein <- all.dat %>% 
  group_by(ORG_EIN) %>% 
  summarise(count = n()) %>% 
  filter(count > 1)

all.dat <-
  all.dat %>% 
  filter(!(ORG_EIN %in% remove.ein))


#merge
dat.2018 <-
  dat.2018 %>% 
  select(NTEECC:hosp, ntee2.code, EIN, LONGITUDE, LATITUDE, STATE, ZIP5, FIPS) %>% 
  filter(EIN %in% features2_ID$ORG_EIN) %>%
  right_join(all.dat, by = join_by("EIN" == "ORG_EIN"))


```

### Geographic

```{r}
### By State
dat.2018 %>%
  select(STATE, total.score) %>%
  na.omit() %>%
  group_by(STATE) %>% 
  summarise(AvgScore = mean(total.score),
            SDScore = sd(total.score),
            count = n())  

states_sf <- get_urbn_map(map = "states", sf = TRUE)

state.dat <- dat.2018 %>%
  group_by(STATE) %>% 
  summarise(AvgScore = mean(total.score))

states_sf %>% 
  left_join(state.dat, by = join_by("state_abbv" == "STATE")) %>%
  ggplot() +
  geom_sf(mapping = aes(fill = AvgScore)) 


## by County 
counties_sf <- get_urbn_map(map = "counties", sf = TRUE)

counties.dat <- dat.2018 %>%
  group_by(FIPS) %>% 
  summarise(AvgScore = mean(total.score))   %>% 
  mutate(FIPS = str_pad(FIPS, 5, pad = "0"))

counties_sf %>% 
  left_join(counties.dat, by = join_by("county_fips" == "FIPS")) %>%
  ggplot() +
  geom_sf(mapping = aes(fill = AvgScore))


```

### NTEE

```{r}
#by broad category
dat.2018 %>%
  select(ntee2.code, total.score) %>%
  na.omit() %>%
  mutate(bc = substr(ntee2.code, 1, 3)) %>% 
  group_by(bc) %>% 
  summarise(AvgScore = mean(total.score),
            SDScore = sd(total.score),
            count = n()) 

dat.2018 %>%
  select(ntee2.code, total.score) %>%
  na.omit() %>%  
  mutate(bc = substr(ntee2.code, 1, 3)) %>% 
  ggplot(aes(x = total.score))+
  geom_density(aes(color = bc)) +
  ggtitle("Total Score by Broad Category")


## By org type 
dat.2018 %>%
  select(ntee2.code, total.score) %>%
  na.omit() %>%
  mutate(org.type = substr(ntee2.code, 9, 10)) %>% 
  group_by(org.type) %>% 
  summarise(AvgScore = mean(total.score),
            SDScore = sd(total.score),
            count = n())  

dat.2018 %>%
  select(ntee2.code, total.score) %>%
  na.omit() %>%
  mutate(org.type = substr(ntee2.code, 9, 10)) %>% 
  ggplot(aes(x = total.score))+
  geom_density(aes(color = org.type)) +
  ggtitle("Total Score by Org Type")


```
